LPS-116842 Extract token component to it's own file
